Add files via upload
With the rise in aerial vehicles, efficient identification is crucial for aviation security, air traffic management, and surveillance. Manual monitoring is slow and error-prone, highlighting the need for an automated system. This project develops an AI-based model to detect and classify aerial vehicles like airplanes and helicopters in real-time, enhancing safety and operational efficiency
Our approach to building this Aerial object classifier model is discussed in four steps:

Explore the dataset
Build a CNN model
Train and validate the model
Test the model with test dataset